# Portabrasil-server/docs/apifox-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PortaBrasil Flask 后端接口",
    "version": "0.1.0",
    "description": "PortaBrasil PDF 单据上传、解析入库、业务查询和费用明细接口。可直接导入 Apifox。"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5001",
      "description": "本地开发服务"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/files/upload": {
      "post": {
        "tags": ["Files"],
        "summary": "上传 PDF 文件",
        "description": "上传 PDF。默认 parse=true，会立即调用智谱 PDF 解析并将结果写入数据库；parse=false 时只保存文件记录。",
        "operationId": "uploadPdfFile",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "PDF 文件"
                  },
                  "parse": {
                    "type": "boolean",
                    "default": true,
                    "description": "是否上传后立即解析入库"
                  }
                }
              },
              "encoding": {
                "file": {
                  "contentType": "application/pdf"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "重复文件或已有文件返回",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                }
              }
            }
          },
          "201": {
            "description": "文件上传成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                },
                "examples": {
                  "uploadedWithoutParse": {
                    "summary": "parse=false",
                    "value": {
                      "duplicated": false,
                      "file": {
                        "id": 1,
                        "file_name": "p2s.pdf",
                        "file_path": "/path/to/uploads/20260415/xxx.pdf",
                        "file_size": 155514,
                        "file_hash": "sha256hex",
                        "page_count": 1,
                        "parse_status": "PENDING",
                        "upload_time": "2026-04-15 22:00:00",
                        "remark": null
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/files/{file_id}/parse": {
      "post": {
        "tags": ["Files"],
        "summary": "解析已有 PDF 文件",
        "description": "对已上传的 PDF 文件发起解析任务，解析成功后写入 customs_business 和 customs_business_fee_item。",
        "operationId": "parseExistingFile",
        "parameters": [
          {
            "$ref": "#/components/parameters/FileId"
          }
        ],
        "responses": {
          "200": {
            "description": "解析成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParseFileResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/files": {
      "get": {
        "tags": ["Files"],
        "summary": "查询 PDF 文件列表",
        "description": "分页查询上传过的 PDF 文件，可按解析状态筛选。",
        "operationId": "listFiles",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "每页数量，最大 100",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "description": "偏移量",
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "解析状态",
            "schema": {
              "$ref": "#/components/schemas/ParseStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "文件列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/{file_id}": {
      "get": {
        "tags": ["Files"],
        "summary": "查询 PDF 文件详情",
        "description": "查询 PDF 文件记录、关联解析任务和来源业务记录。",
        "operationId": "getFileDetail",
        "parameters": [
          {
            "$ref": "#/components/parameters/FileId"
          }
        ],
        "responses": {
          "200": {
            "description": "文件详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileDetailResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "在请求头中添加：Authorization: Bearer <access_token>"
      }
    },
    "parameters": {
      "FileId": {
        "name": "file_id",
        "in": "path",
        "required": true,
        "description": "pdf_file.id",
        "schema": {
          "type": "integer",
          "example": 1
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "请求参数错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "badRequest": {
                "value": {
                  "error": "缺少 multipart 字段 file"
                }
              }
            }
          }
        }
      },
...
      "NotFound": {
        "description": "资源不存在",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "notFound": {
                "value": {
                  "error": "业务记录不存在"
                }
              }
            }
          }
        }
      },
      "ServerError": {
        "description": "服务端错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "missingKey": {
                "value": {
                  "error": "ZHIPU_API_KEY 未配置，无法调用智谱 PDF 解析接口"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "ParseStatus": {
        "type": "string",
        "enum": ["PENDING", "PROCESSING", "SUCCESS", "FAILED"],
        "example": "SUCCESS"
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "错误信息"
          }
        },
        "required": ["error"]
      },
      "PdfFile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "file_name": {
            "type": "string",
            "example": "p2s.pdf"
          },
          "file_path": {
            "type": "string",
            "example": "/Users/leo/Documents/CODE/PortaBrasil/Portabrasil-server/uploads/20260415/xxx.pdf"
          },
          "file_size": {
            "type": "integer",
            "example": 155514
          },
          "file_hash": {
            "type": "string",
            "example": "3c5f2f0d2b7b..."
          },
          "page_count": {
            "type": "integer",
            "nullable": true,
            "example": 1
          },
          "parse_status": {
            "$ref": "#/components/schemas/ParseStatus"
          },
          "upload_time": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-15 22:00:00"
          },
          "remark": {
            "type": "string",
            "nullable": true,
            "example": null
          }
        }
      },
      "ParseTask": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "file_id": {
            "type": "integer",
            "example": 1
          },
          "task_no": {
            "type": "string",
            "example": "2f6b0f5f8bbf43f39f410e01c12a99c7"
          },
          "parser_type": {
            "type": "string",
            "example": "LLM"
          },
          "status": {
            "$ref": "#/components/schemas/ParseStatus"
          },
          "start_time": {
            "type": "string",
            "nullable": true,
            "example": "2026-04-15 22:01:00"
          },
          "end_time": {
            "type": "string",
            "nullable": true,
            "example": "2026-04-15 22:01:12"
          },
          "error_message": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "raw_result": {
            "type": "string",
            "nullable": true,
            "description": "智谱原始解析结果 JSON 字符串或错误信息",
            "example": "{\"status\":\"succeeded\",\"content\":\"...\"}"
          },
          "created_time": {
            "type": "string",
            "example": "2026-04-15 22:01:00"
          },
          "updated_time": {
            "type": "string",
            "example": "2026-04-15 22:01:12"
          }
        }
      },
      "CustomsBusiness": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "s_ref": {
            "type": "string",
            "example": "20250528000158"
          },
          "n_ref": {
            "type": "string",
            "nullable": true,
            "example": "2-12626/25"
          },
          "document_no": {
            "type": "string",
            "nullable": true,
            "example": "12626"
          },
          "invoice_no": {
            "type": "string",
            "nullable": true,
            "example": "20250528000158"
          },
          "nf_no": {
            "type": "string",
            "nullable": true,
            "example": "000000000014069"
          },
          "process_no": {
            "type": "string",
            "nullable": true,
            "example": "2-12626/25"
          },
          "business_type": {
            "type": "string",
            "nullable": true,
            "example": "IMPORT"
          },
          "trade_company": {
            "type": "string",
            "nullable": true,
            "example": "HAOY TRADING"
          },
          "customer_name": {
            "type": "string",
            "nullable": true,
            "example": "LAIOUN'S IMPORTACAO E EXPORTACAO LTDA ME"
          },
          "customer_address": {
            "type": "string",
            "nullable": true,
            "example": "RUA MARIA RODRIGUES No.:164"
          },
          "customer_city": {
            "type": "string",
            "nullable": true,
            "example": "RIO DE JANEIRO"
          },
          "customer_state": {
            "type": "string",
            "nullable": true,
            "example": "RJ"
          },
          "customer_zip_code": {
            "type": "string",
            "nullable": true,
            "example": "21031-490"
          },
          "customer_tax_no": {
            "type": "string",
            "nullable": true,
            "example": "09.089.839/0001-12"
          },
          "issuer_name": {
            "type": "string",
            "nullable": true,
            "example": "LOGIMEX COMERCIO EXTERIOR LTDA"
          },
          "issuer_tax_no": {
            "type": "string",
            "nullable": true,
            "example": "02.155.413/0001-70"
          },
          "mawb_mbl": {
            "type": "string",
            "nullable": true,
            "example": "YMJAB232097847"
          },
          "hawb_hbl": {
            "type": "string",
            "nullable": true,
            "example": "YMJAB232097847"
          },
          "di_duimp_due": {
            "type": "string",
            "nullable": true,
            "example": "25/1439746-3"
          },
          "registration_date": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "2025-07-01"
          },
          "arrival_date": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "2025-06-25"
          },
          "customs_clearance_date": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "2025-07-01"
          },
          "loading_date": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "2025-07-03"
          },
          "destination": {
            "type": "string",
            "nullable": true,
            "example": "RIO DE JANEIRO"
          },
          "vessel_flight_name": {
            "type": "string",
            "nullable": true,
            "example": "EVER FUTURE"
          },
          "gross_weight": {
            "type": "string",
            "nullable": true,
            "description": "Decimal 以字符串返回",
            "example": "15280.000"
          },
          "volume_count": {
            "type": "integer",
            "nullable": true,
            "example": 191
          },
          "cargo_desc": {
            "type": "string",
            "nullable": true,
            "example": "X12-10"
          },
          "freight_currency": {
            "type": "string",
            "nullable": true,
            "example": "USD"
          },
          "freight_amount": {
            "type": "string",
            "nullable": true,
            "example": "2284.71"
          },
          "fob_currency": {
            "type": "string",
            "nullable": true,
            "example": "USD"
          },
          "fob_amount": {
            "type": "string",
            "nullable": true,
            "example": "45656.29"
          },
          "cif_currency": {
            "type": "string",
            "nullable": true,
            "example": "USD"
          },
          "cif_amount": {
            "type": "string",
            "nullable": true,
            "example": "47941.00"
          },
          "cif_brl_amount": {
            "type": "string",
            "nullable": true,
            "example": "261618.80"
          },
          "dollar_rate": {
            "type": "string",
            "nullable": true,
            "example": "5.4571000"
          },
          "euro_rate": {
            "type": "string",
            "nullable": true,
            "example": "0.0000000"
          },
          "total_credit": {
            "type": "string",
            "nullable": true,
            "example": "228389.13"
          },
          "total_debit": {
            "type": "string",
            "nullable": true,
            "example": "200897.37"
          },
          "balance_amount": {
            "type": "string",
            "nullable": true,
            "example": "27491.76"
          },
          "balance_direction": {
            "type": "string",
            "nullable": true,
            "example": "FAVOR"
          },
          "source_file_id": {
            "type": "integer",
            "nullable": true,
            "example": 1
          },
          "source_page_no": {
            "type": "integer",
            "nullable": true,
            "example": 1
          },
          "data_status": {
            "type": "string",
            "example": "DRAFT"
          },
          "created_time": {
            "type": "string",
            "example": "2026-04-15 22:02:00"
          },
          "updated_time": {
            "type": "string",
            "example": "2026-04-15 22:02:00"
          }
        }
      },
      "FeeItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "business_id": {
            "type": "integer",
            "example": 1
          },
          "fee_date": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "2025-06-26"
          },
          "fee_code": {
            "type": "string",
            "nullable": true,
            "example": "112"
          },
          "fee_name": {
            "type": "string",
            "example": "CAPATAZIA"
          },
          "credit_amount": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "debit_amount": {
            "type": "string",
            "nullable": true,
            "example": "600.00"
          },
          "line_no": {
            "type": "integer",
            "nullable": true,
            "example": 1
          },
          "raw_text": {
            "type": "string",
            "nullable": true,
            "example": "26.06.2025 112 CAPATAZIA 600,00"
          },
          "created_time": {
            "type": "string",
            "example": "2026-04-15 22:02:00"
          }
        }
      },
      "CustomsBusinessWithFees": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CustomsBusiness"
          },
          {
            "type": "object",
            "properties": {
              "fee_items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/FeeItem"
                }
              }
            }
          }
        ]
      },
...
      "FileUploadResponse": {
        "type": "object",
        "properties": {
          "duplicated": {
            "type": "boolean",
            "example": false
          },
          "file": {
            "$ref": "#/components/schemas/PdfFile"
          },
          "task_id": {
            "type": "integer",
            "nullable": true,
            "example": 1
          },
          "business": {
            "$ref": "#/components/schemas/CustomsBusinessWithFees"
          }
        }
      },
      "ParseFileResponse": {
        "type": "object",
        "properties": {
          "file": {
            "$ref": "#/components/schemas/PdfFile"
          },
          "task_id": {
            "type": "integer",
            "example": 1
          },
          "business": {
            "$ref": "#/components/schemas/CustomsBusinessWithFees"
          }
        }
      },
      "FileListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PdfFile"
            }
          },
          "total": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 20
          },
          "offset": {
            "type": "integer",
            "example": 0
          }
        }
      },
      "FileDetailResponse": {
        "type": "object",
        "properties": {
          "file": {
            "$ref": "#/components/schemas/PdfFile"
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParseTask"
            }
          },
          "businesses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomsBusiness"
            }
          }
        }
      },
...
    }
  }
}